Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-  float myFloatNum;
-  scanf("%f", &myFloatNum);
+  double myFloatNum;
+  scanf("%lf", &myFloatNum);
   printf("You entered: %.4f", myFloatNum); 
   return 0;
 }
